error checking and logging when loading .bib files
loading a .bib file used to be a pretty silent event. now, ebib has a log
buffer into which information about the loading is logged. warnings about
unknown or duplicate entries and errors (for the moment just unbalanced
delimiters) are also reported here. when warnings or errors occur, the user
is given a message with the suggestion to check out the log.