Move url-splitting code to separate function.
This fixes a bug where < and > in DOIs weren't properly handled: since the
contents of the 'doi' field was parsed by the url-splitting code in
ebib-browse-url, any < and > in the DOI were taken as delimiting a separate
URL. The 'doi' field can only contain one DOI, however, so there is no need
to split it.

Thanks to Elias Pipping for reporting this bug.