ebib-db-get-field-value: check for alias after checking for cross-ref.
This prevents a bug when a field has an alias but the alias has no
value. If the cross-ref is checked afterwards, there is a value and
since an alias exists, the return string is marked as being from an
alias field (in addition to being marked as coming from a
cross-referenced field).