When (re)starting Ebib, only update the buffers if there is a need.
When calling `M-x ebib', the buffers only need to be updated if Ebis was
started for the first time, or when a file or a key argument were
passed. Otherwise, the contents of the buffers is up to date.

This speeds up restarting Ebib when a lot of entries are loaded.