Allow command-line options in ebib-file-associations.
Previously, the viewer name in ebib-file-associations needed to be
precisely an executable to be found on exec-path. Now it can also
contain command-line options, and a "%s" where the file path needs to
go.

If the string contains no "%s", it falls-back to the original behavior,
and is therefore backwards compatible.