Update for parsebib v. 2.3
In order to support @comments that are not surrounded by braces, parsebib 2.3
returns @comments *with* braces if they are present. This has to be accommodated
for.

For consistency's sake, parsebib-read-preamble now also returns the braces
around the preamble text, which also needs to be accounted for.

Version number is updated to 2.11.