`ebib--update-index-buffer`: always set the current entry.
Switching to the index buffer without updating its contents would not change
point, even if the current entry of the database was changed programmatically.
Now, switching to the index buffer always resets the current entry.

This makes it possible to call `ebib--update-buffers` with a `no-refresh`
argument in `ebib`, so that calling `M-x ebib` on an entry key in e.g., a LaTeX
buffer is much quicker now.