Use completion when editing certain field values.
Completion is offered for the author and editor fields, and for the journal,
journaltitle, publisher and organization fields.

Gathering a collection for the author and editor fields can be somewhat slow,
there is an option to disable it. Note that completion for these fields is only
offered when they are empty.