Bug fix: do not update index buffer when Ebib is not active.
Some commands, such as `ebib-import` and `ebib-insert-citation` can modify a
database when Ebib is not in the foreground. In such cases, the affected
database's index buffer needs to be updated. Before, this was done with a call
to `ebib-update-index-buffer`, but this is not correct, since it updates the
index buffer of `ebib--cur-db`, but `ebib-insert-citation` may actually modify a
different database, which would then not get its index buffer updated.

* (ebib-db-kill-buffer): removed. Killing the index buffer is not the right way
  to indicate that it needs to be updated, because it leads to problems when
  returning to Ebib with `M-x ebib`: if the index buffer of `ebib--cur-db` is
  killed, an error is raised (Attempt to display dead buffer).
* ebib--dirty-index-buffer: new variable (buffer-local). Set to t if an index
  buffer needs to be updated.