Add raw.lua filter.
This filter converts {{raw}} and {{endraw}} directives into {% raw %} and {%
endraw %} in the gfm output and removes them in all other output
formats (currently only texinfo). This is needed because the manual is published
on Github, which uses Jekyll, which uses the Liquid templating language. Liquid
uses directives of the form {% ... %}, but unfortunately the Ebib manual
contains a few occurrences of {% ... %} that are not Liquid directives. They
need to be wrapped in {% raw/endraw %} directives in order to tell Liquid/Jekyll
to treat them as normal text.

This commit undoes the pre-previous commit where the same thing is done with
pandoc-mode directives.