Remove `ebib--select-entry`.
`ebib--select-entry` is not needed anymore, because the cursor position in the
index buffer already determines what the current entry key is. If the entry key
is not the first column in index buffer, it would produce incorrect results,
which in most cases was masked by a subsequent call to
`ebib--update-entry-buffer`, so the bug was never discovered.