Bug fix: ensure that an index buffers is a newly created buffer.
This fixes a long-standing bug that I never managed to pin down before. (See,
e.g., Github issue #147).

Turns out that if a buffer is visiting a `.bib` file and you then open that same
`.bib` file in Ebib, the buffer visiting that `.bib` file would get reused as
the index buffer in Ebib. Since this buffer is still visiting the `.bib` file,
the contents of the index buffer may at some point get written to the file.