Enable adding the bib file to Org links.
Org links are supposed to be *links*, not citations, so it makes sense that you
can specify exactly where the bib entry is to be found. Since there are other
ways to do this, it shouldn't be the only option, but it should be an option.

This is also in preparation for the upcoming citation support in Org mode.