Only write filtered entries when writing to a new file.
If a filtered database is written to a new file (`ebib-write-database`), only
the filtered entries should be written, but this was not the case before.

Replace call to `ebib-db-list-keys` with a call to `ebib--list-keys` in
`ebib--format-database-as-bibtex`. Also move `ebib--format-database-as-bibtex`
to `ebib.el` in order to avoid a circular dependency (`ebib--list-keys` is
define in `ebib.el`).

This fixes part of Github issue #174.