`ebib-add-field`: Use empty string as initial value.
`ebib--list-undefined-fields` strips out fields without a value, so in order to
make the new field visible, it needs a non-nil value.

Editing the newly added field is now done in a `condition-case` to catch an
explicit `C-g` by the user and remove the field again.

This should fix Github issue #176.