Use more conventional key bindings in multiline edit buffer.
The key bindings `C-c C-c` / `C-c C-k` are conventionally used in various
Emacs packages for finalising / cancelling a temporary buffer; cf. e.g., Org
capture, `message-mode`, VC (apparently), but also external packages such as
magit. It makes sense to use them for the multiline edit buffer as well.

A comment is added to the manual to explain how to return to the original key
bindings using `C-c |` as a prefix.