Remove indentation spaces in multiline fields
When bibtex entries are created, some fields have multiple lines that are indented for readability in the *.bib files. Unfortunately, those spaces, which appear only on the 2nd and later lines, get inserted into the database. This patch eliminates those leading spaces before adding the 19 spaces that are used for displaying multiline fields, making the indentation right.