ebib-jump-to-field: add ivy action to kill field contents
Add ivy action on 'k' in ebib-jump-to-field to kill field's contents.

Redefine ebib-kill-field-contents to take a field name and kill that
field's contents in the current entry. Reimplement the old
behaviour (killing the contents of the field at point) in
ebib-kill-current-field. Replace the old function name with new one to
retain functionality throughout the code.