ebib-jump-to-field: add ivy action to delete field contents
Add ivy action on 'd' in ebib-jump-to-field to delete field's
contents, without adding them to the kill ring.

Redefine ebib-delete-field-contents to take a field name and delete that
field's contents in the current entry. Reimplement the old
behaviour (deleting the contents of the field at point) in
ebib-delete-current-field. Replace the old function name with new one to
retain functionality throughout the code.