ebib-jump-to-field: add ivy action to copy field contents
Add ivy action on 'c' in ebib-jump-to-field to copy field's contents.

To do this, generalise ebib-copy-field-contents to take the name of
the field to copy from. Subsequently, reimplement the original
functionality (always copying from the field at point) in a new
function, ebib-copy-current-field-contents. Replace the old function
name with new one to retain functionality throughout the code.