ebib-jump-to-field: add ivy action to view field contents as help
Add ivy action on 'v' in ebib-jump-to-field to view field in a help
buffer.

Redefine ebib-view-field-as-help to take a field name and view that
field in the current entry. Reimplement the old behaviour (view the
field at point) in ebib-view-current-field-as-help. Replace the old
function name with new one where appropriate to retain functionality
throughout the code.