ebib-jump-to-field: add ivy action to edit field contents
Add ivy action on 'e' in ebib-jump-to-field to edit field's
contents.

Redefine ebib-edit-field-contents to take a field name and edit that
field's contents in the current entry. Reimplement the old
behaviour (edit the contents of the field at point) in
ebib-edit-current-field. Replace the old function name with new one
where appropriate to retain functionality throughout the code.