ebib-jump-to-field: add ivy action to toggle special status
Add ivy action on 'r' in ebib-jump-to-field to toggle whether a field
is stored unbraced or not.

Redefine ebib-toggle-raw to take a field name and toggle that field in
the current entry. Reimplement the old behaviour (toggle the field at
point) in ebib-toggle-raw-current-field. Replace the old function name
with new one where appropriate to retain functionality throughout the
code.