ebib-jump-to-field: add ivy action to browse url
Add ivy action on 'u' in ebib-jump-to-field to browse url of current
entry.

Since this calls ebib-browse-url from lisp, edit the function to get
the prefix arg in a way which does not rely on its being called
interactively.

The ivy api assumes that every function (second element) in the action
list will be callable with a single string arg: the selected
candidate. Since this is not true of ebib-browse-url, wrap it in a
lambda with a single null argument.