ebib-jump-to-field: add ivy action to browse doi
Add ivy action on 'I' in ebib-jump-to-field to browse doi of current
entry.

The ivy api assumes that every function (second element) in the action
list will be callable with a single string arg: the selected
candidate. Since this is not true of ebib-browse-doi, wrap it in a
lambda with a single null argument.