ebib-jump-to-field: add ivy action to yank field contents
Add ivy action on 'y' in ebib-jump-to-field to yank field's contents.

Redefine ebib-yank-field-contents to take a field name and yank that
field's contents in the current entry. Reimplement the old
behaviour (yanking the contents of the field at point) in
ebib-yank-current-field. Replace the old function name with new one to
retain functionality throughout the code.