ebib-jump-to-field: add ivy action to edit multiline field contents
Add ivy action on 'm' in ebib-jump-to-field to edit field's
contents as multiline.

This caused an error with ebib-jump-to-field, where re-search-forward
would look for the name of the field in the multiline editing buffer,
so suppress errors in that function from re-search-forward, and only
goto the relevant char if it is non-nil.