ebib-index-columns: last value can specify custom sort function
If the last value of a member of ebib-index-columns is non-nil, that
column of the index can be used for sorting. If it is t, comparison
for the sort is done with string-collate-lessp. Otherwise, the symbol
which is the third value of the member is called as a function.

Update docstrings of ebib--sort-keys-list and ebib-index-columns to
reflect this.