ebib--sort-keys-list: guess appropriate sorting methods per field
New option: ebib-field-sort-functions-alist, an alist of
case-insensitive field names and the functions for sorting them. So
far this is only used for integer fields, with
ebib-compare-numerical-strings.

New function: ebib-compare-numerical-strings, for comparing which of
two strings represents a higher value.

Within ebib--sort-keys-list, if the last value in a field's entry in
ebib-index-columns is t, check if it has a corresponding sorting
function in ebib-field-sort-functions-alist before defaulting to
string-collate-lessp