Only display hidden and undefined fields that actually have a value.
When a field value is killed, its entry is not removed from the alist
representing the BibTeX entry; rather, its value is set to `nil`. Before this
commit, this meant that they are still displayed in the entry buffer, which is
confusing if the field is a hidden or undefined field.