Allow user to customise function to extract note text
new custom variable: ebib-notes-extract-text-function. Points to a
function which should return a list of strings in the note of the
given key. ebib--extract-note-text is now just a wrapper over calling
this function. ebib-notes-extract-text-function defaults to...

new function: ebib-extract-note-text-default. Unifies the work
previously done by ebib--extract-note-text and
ebib--extract-note-text-1 into one backend function. Has sensible
defaults for notes in separate files or the same file.