ebib-delete-field-contents: support deleting/killing external notes
Trying to delete the field 'external' (displayed as 'external note')
will display appropriate message ('field' is replaced by 'note'). When
`ebib-notes-storage' is set to `one-file-per-note', deletion is
affected by deleting the file -- otherwise it is not supported and an
error to this effect is raised, suggesting the user delete the note
themselves. When the `kill' argument is non-nil, the full contents of
the note (as returned by `(ebib--extract-note-text key nil)') are
pushed to the kill ring.