Vary return value of `ebib--bib-find-bibtex-entries`.
Add an optional argument `list-entries` to `ebib--bib-find-bibtex-entries`
which, if `true`, causes `ebib--bib-find-bibtex-entries` to return a list of
entry keys instead of the summary count.

An optional argument is also added to `ebib-import-entries`, which has basically
the same effect.

The reason for this change is Github PR #241. If `ebib--bib-find-bibtex-entries`
reads a large `.bib` file, we don't want to create a long list of entry keys.
When reading only a few or even one single entry, having the entry key can be
useful.