ebib-clean-TeX-markup: generalise + use regexps
Match properly on regexps, making ebib-TeX-markup-replace-alist a more
powerful variable.

Using proper regexps and replace-regexp-in-string means that the
replacement can now be a function which generates the relevant text.
This obviates the need for the hard-coded specific cases previously in
the function. Move these to (regexp . lambda) pairs in the variable.

Update docstrings to reflect changes.