ebib--edit-field-as-multiline: align with `ebib-field-edit-functions'
ebib--edit-field-as-multiline: add new argument INIT-CONTENTS, and
make all arguments except FIELD optional. This allows
`ebib--edit-field-as-multiline' to be used as an ivy action. Ignore
INIT-CONTENTS (so rename to _INIT-CONTENTS). This ensures if the
actual contents of the field are unbraced, then these are what is
tested in the function (since `ebib-edit-field' unbraces contents
before passing them to this function).

ebib-edit-field: update to ignore init-contents.

ebib-field-edit-functions: update to use
`ebib--edit-field-as-multiline' for multiline fields.

ebib--edit-multiline-field: remove, no longer necessary.

ebib ivy: update to use `ebib--edit-field-as-multiline'.