ebib-follow-crossref: automatically follow current field's crossref
When called from a field which includes crossref keys, only consider
those keys (if it contains one key, jump to it, if more than one,
prompt the user, but only offer the ones in the current field). prefix
arg forces prompting for all keys regardless.

This behaviour can be disabled by setting
`ebib-follow-current-field-crossref' to nil.