Remove `ebib--edit-entry-type`.
With the new field editing functionality, this function is reduced to a single
line, so we can remove it.

This also fixes a bug where editing the entry type would trigger an error saying
`"type" not found`. This was due to `ebib--redisplay-field` being called, which
redisplays the new field and then places point on it using `re-search-forward`.
But because the field name being searched for is `=type=` while the field in the
entry buffer is called `type`, this search would yield an error.