Bugfix: allow typing space in field editing functions.
By default, SPC completes the current entry in `completing-read-multiple`,
rather than insert space. Since most (all?) fields allow spaces, this is not
desirable. To remedy this, I unbound SPC in `crm-local-must-match-map`, but this
is not correct in calls to `completing-read-multiple` that do not require a
match. In relevant instances, the unbinding of SPC must be done in `crm-local-completion-map`.