Do not kill notes buffer for `one-file-per-note`.
Until now, with the setting `one-file-per-note` for `ebib-notes-storage`, every
time the entry buffer is updated, the notes buffer is killed. This was mainly to
prevent dozens of notes buffers from staying alive in an Emacs session. However,
it leads to a problem with the setting `all` for `ebib-notes-show-note-method`,
as seen in Github issue #283.

We may decide to turn this into a user-configurable option if users are annoyed
by the many note buffers staying open.