Add a column for the mark.
In the old index buffer, a marked entry is displayed with inverse face, but this
is problematic for the vtable, because it doesn't seem possible to apply a face
to the spaces between the data elements.

The idea is therefore to add a separate, nameless, column for a mark sign (by
default an asterisk).

As it turns out, however, there are some bugs in vtable.el, one of which affects
vtable-update-object, which make this strategy impossible at the moment. The
next Emacs version (version 30) should have a fix for it.