ebib-preload-bib-files: autoload safety predicate
This is useful because the variable definition, including the `:safe:'
declaration, is only loaded as the package is loaded. If the user sets
the variable locally (e.g. in `Local Variables') and then visits that
file in a new session before opening ebib for the first time, the safe
predicate won't be recognised and the user is prompted for whether to
use the new value. Using an autoload is the solution to this suggested
in the elisp manual `(elisp) File Local Variables'.