Extract DOI from doi field & allow customization of resolver
This is done to a) prevent oddities if the doi is a whole URL but not
using dx.doi.org as the resolver, and b) allow for alternative
resolvers (e.g., https://oadoi.org/) to be used.

Additionally, it switches the default resolver to `https://doi.org`,
which is the DOI Foundation's preferred default resolver.