Improve TeX clean-up.
Note creation would reuse the functions used for displaying BibTeX entries
in the entry buffer and with `ebib-jump-to-entry`. This would remove any
sign of font commands \textit, \textbf, \textsc and \emph. This update,
together with the update to parsebib 5.0, ensures that when creating a note
in Org format, those commands are rendered more appropriately using Org
markup.

Additionally, add functions for rendering field values using Markdown.